+++ lib/view_model.dart
@@ -2,5 +2,4 @@ library view_model;
 
 export "package:view_model/src/view_model/config.dart";
 export "package:view_model/src/view_model/extension.dart";
-export "package:view_model/src/view_model/stateless_view_model.dart";
 export "package:view_model/src/view_model/view_model.dart";