+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.4.0
+
+_Breaking change:_
+
+* Use `ViewModel` instead of `StatelessViewModel`.
+* Use `StateViewModel` instead of `ViewModel`.
+* Use either `watchViewModel` or `readViewModel` instead of `getViewModel`/
+  `requireExistingViewModel`.
+* Use `StateViewModel.listenState` instead of `ViewModel.listen`.
+* Use `ViewModel.listen` instead of `ViewModel.addListener`.
+
+* Support `ViewModel.read<T>` to read existing view model globally.
+
 ## 0.3.0
-* transfer to https://github.com/lwj1994/flutter_view_model. thank to [Miolin](https://github.com/Miolin)
+
+* transfer to https://github.com/lwj1994/flutter_view_model. thank
+  to [Miolin](https://github.com/Miolin)